Server-NegevNerds/requirements.txt
alembic==1.14.0
appnope==0.1.4
asttokens==2.4.1
bcrypt==4.2.1
beautifulsoup4==4.12.3
black==24.8.0
blinker==1.9.0
certifi==2024.8.30
cffi==1.17.1
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
comm==0.2.2
contourpy==1.2.1
cryptography==44.0.0
cycler==0.12.1
debugpy==1.8.5
decorator==5.1.1
distlib==0.3.8
distro==1.9.0
et_xmlfile==2.0.0
executing==2.0.1
filelock==3.15.4
filterpy==1.4.5
Flask==3.1.0
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
fonttools==4.53.1
identify==2.6.0
idna==3.10
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.26.0
isort==5.13.2
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
Mako==1.3.8
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mypy-extensions==1.0.0
nest-asyncio==1.6.0
nodeenv==1.9.1
numpy==2.1.0
opencv-python==4.10.0.84
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
parso==0.8.4
pathspec==0.12.1
pdfminer.six==20221105
pdfplumber==0.9.0
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
pluggy==1.5.0
pre-commit==3.8.0
prompt_toolkit==3.0.47
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pycryptodome==3.21.0
Pygments==2.18.0
PyJWT==2.10.1
PyMuPDF==1.25.1
pyparsing==3.1.4
PyPDF2==1.28.6
pypdfium2==4.30.0
pytesseract==0.3.13
pytest==8.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
requests==2.32.3
scikit-learn==1.5.1
scipy==1.14.1
six==1.16.0
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.36
stack-data==0.6.3
tabula-py==2.10.0
threadpoolctl==3.5.0
tika==2.6.0
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
virtualenv==20.26.3
Wand==0.6.13
wcwidth==0.2.13
Werkzeug==3.1.3
